Added basic decision for proving boolean actions
